Bunch of improvements
